Support LDFLAGS in configure and Makefile.config.in except for flexlink.
OC_LDFLAGS is for options specific to linking OCaml programs, LDFLAGS
are user-chosen flags that should be included in any call to the
system linker.

This is intended to fix #9191.

The logic followed in this PR is similar to the treatment of CFLAGS in
the build system, which comes from #9837.

Passing LDFLAGS to flexlink is not supported. It is a pain to do
correctly, and we haven't got requests from Windows users yet, only
from users of unix variants.